requirements.txt
openai-whisper
fastapi
uvicorn
mangum
aiohttp
pydantic
moviepy
mutagen
Pillow==8.2.0
google-api-python-client
google-auth-oauthlib 
google-auth-httplib2
oauth2client